chore: relocate validate_tracker_id_to_images + clarify doctring
